Changes revenue type from int to long because of the risk of overflow (for movies like Avatar)

+++ app/src/main/java/com/bogdanorzea/popularmovies/model/objects/Movie.java
@@ -39,7 +39,7 @@ public class Movie {
     @Json(name = "release_date")
     public String releaseDate;
     @Json(name = "revenue")
-    public int revenue;
+    public long revenue;
     @Json(name = "runtime")
     public int runtime;
     @Json(name = "spoken_languages")
